--- backend/adsystem/core/models.py
@@ -1,13 +1,0 @@
-from adsystem.core.Masters.models import MasterModel
-from adsystem.core.roles.models import RoleModel
-from adsystem.core.users.models import UserModel
-from adsystem.core.permissions.models import PermissionModel, PermissionAction
-
-__all__ = [
-    'UserModel',
-    'RoleModel',
-    'PermissionModel',
-    'PermissionAction',
-    'MasterModel',
-
-]
