+++ backend/ipproject/core/models.py
@@ -1,0 +1,13 @@
+from ipproject.core.Masters.models import MasterModel
+from ipproject.core.roles.models import RoleModel
+from ipproject.core.users.models import UserModel
+from ipproject.core.permissions.models import PermissionModel, PermissionAction
+
+__all__ = [
+    'UserModel',
+    'RoleModel',
+    'PermissionModel',
+    'PermissionAction',
+    'MasterModel',
+
+]
