+++ backend/ipproject/core/routes.py
@@ -1,9 +1,9 @@
 from starlette.routing import Route, Mount
 from starlette.responses import JSONResponse
 
-from adsystem.core.users.resources import routes as user_routes
-from adsystem.core.roles.resources import routes as roles_routes
-from adsystem.core.permissions.resources import get_apps
+from ipproject.core.users.resources import routes as user_routes
+from ipproject.core.roles.resources import routes as roles_routes
+from ipproject.core.permissions.resources import get_apps
 
 
 async def ping(request):
